adding new workflow from the admin panel
