+++ admin/src/pages/SettingPage/index.tsx
@@ -1,7 +1,6 @@
 import { BaseHeaderLayout, Link, Stack, Typography } from '@strapi/design-system';
 import { CheckPagePermissions, useFetchClient } from '@strapi/helper-plugin';
 import PageWrapper from '../../components/PageWrapper';
-import { TextInput } from '@strapi/design-system';
 import useFetch from '../../hooks/useFetch';
 import useFormattedLabel from '../../hooks/useFormattedLabel';
 import pluginPermissions from '../../permissions';
@@ -12,7 +11,7 @@ import ConfigsTable from '../../components/ConfigsTable';
 export default function ProtectedPage() {
   return (
     <CheckPagePermissions permissions={pluginPermissions.settings}>
-      <SettingPage />
+        <SettingPage />
     </CheckPagePermissions>
   );
 }
@@ -28,25 +27,6 @@ const SettingPage = () => {
   const HEADER_TITLE = useFormattedLabel('settings.headers.title');
   const HEADER_SUBTITLE = useFormattedLabel('settings.headers.subtitle');
 
-  const GITHUB_TOKEN = useFormattedLabel('settings.fields.githubtoken');
-  const REPO = useFormattedLabel('settings.fields.repo');
-  const WORKFLOWID = useFormattedLabel('settings.fields.workflowid');
-  const OWNER = useFormattedLabel('settings.fields.owner');
-  const BRANCH = useFormattedLabel('settings.fields.branch');
-
-  const HINT_GITHUB_TOKEN = useFormattedLabel('settings.fields.hint.githubtoken');
-  const HINT_OWNER = useFormattedLabel('settings.fields.hint.owner');
-  const HINT_REPO = useFormattedLabel('settings.fields.hint.repo');
-  const HINT_WORKFLOWID = useFormattedLabel('settings.fields.hint.workflowid');
-  const HINT_BRANCH = useFormattedLabel('settings.fields.hint.branch');
-
-  const PLACEHOLDER_GITHUB_TOKEN = useFormattedLabel('settings.fields.placeholder.githubtoken');
-  const PLACEHOLDER_OWNER = useFormattedLabel('settings.fields.placeholder.owner');
-  const PLACEHOLDER_REPO = useFormattedLabel('settings.fields.placeholder.repo');
-  const PLACEHOLDER_WORKFLOWID = useFormattedLabel('settings.fields.placeholder.workflowid');
-  const PLACEHOLDER_BRANCH = useFormattedLabel('settings.fields.placeholder.branch');
-  const BUTTON_DETAILS = useFormattedLabel('button.details');
-
   const setConfig = (data: Config) => {
     post(`/${pluginId}/config`, data);
   };
@@ -58,68 +38,6 @@ const SettingPage = () => {
       pageTitle={PAGE_TITLE}
     >
       <ConfigsTable />
-      {/* <Stack spacing={6}>
-        <TextInput
-          type="password"
-          label={GITHUB_TOKEN}
-          aria-label={GITHUB_TOKEN}
-          name="githubToken"
-          placeholder={PLACEHOLDER_GITHUB_TOKEN}
-          defaultValue={githubToken}
-          required
-          HintMessage={
-            <Typography variant="omega">
-              {HINT_GITHUB_TOKEN}{' '}
-              <Link
-                href="https://docs.github.com/en/authentication/keeping-your-account-and-data-secure/creating-a-personal-access-token"
-                isExternal
-              >
-                {BUTTON_DETAILS}
-              </Link>
-            </Typography>
-          }
-        />
-        <TextInput
-          type="text"
-          label={OWNER}
-          aria-label={OWNER}
-          name="githubAccount"
-          placeholder={PLACEHOLDER_OWNER}
-          defaultValue={githubAccount}
-          required
-          HintMessage={<Typography variant="omega">{HINT_OWNER}</Typography>}
-        />
-        <TextInput
-          type="text"
-          label={REPO}
-          aria-label={REPO}
-          name="repo"
-          placeholder={PLACEHOLDER_REPO}
-          defaultValue={repo}
-          required
-          HintMessage={<Typography variant="omega">{HINT_REPO}</Typography>}
-        />
-        <TextInput
-          type="text"
-          label={WORKFLOWID}
-          aria-label={WORKFLOWID}
-          name="workflow_id"
-          placeholder={PLACEHOLDER_WORKFLOWID}
-          defaultValue={workflow}
-          required
-          HintMessage={<Typography variant="omega">{HINT_WORKFLOWID}</Typography>}
-        />
-        <TextInput
-          type="text"
-          label={BRANCH}
-          aria-label={BRANCH}
-          name="branch"
-          placeholder={PLACEHOLDER_BRANCH}
-          defaultValue={branch}
-          required
-          HintMessage={<Typography variant="omega">{HINT_BRANCH}</Typography>}
-        />
-      </Stack> */}
     </PageWrapper>
   );
 };
